+++ internal/server/containerstore.go
@@ -32,6 +32,8 @@ type savedItem struct {
 	Dmg   int      `json:"dmg,omitempty"`
 	Ench  int32    `json:"ench,omitempty"`
 	Ench2 int32    `json:"ench2,omitempty"` // enchantments 3-4 (2026-09-06)
+	Ench3 int32    `json:"ench3,omitempty"` // enchantments 5-6 (2026-09-19)
+	Ench4 int32    `json:"ench4,omitempty"` // enchantments 7-8
 	MapID int32    `json:"map_id,omitempty"`
 	Pats  [6]int32 `json:"pats,omitempty"` // banner layers, patPlus1<<8|color
 	Trim  int32    `json:"trim,omitempty"` // (mat+1)<<8|(pat+1)